+++ packages/shared/src/enums/index.ts
@@ -1,3 +1,2 @@
-export * from './http-method';
 export * from './http-header';
 export * from './http-status';